requirements.txt
# asgiref==3.9.1
# certifi==2025.8.3
# charset-normalizer==3.4.3
# Django==5.2.6
# django-browser-reload==1.20.0
# django-environ==0.12.0
# django-tailwind==4.2.0
# django-widget-tweaks==1.5.0
# idna==3.10
# Jinja2==3.1.6
# MarkupSafe==3.0.2
# mysqlclient==2.2.7
# pillow==11.3.0
# python-dateutil==2.9.0.post0
# python-slugify==8.0.4
# PyYAML==6.0.2
# requests==2.32.5
# six==1.17.0
# sqlparse==0.5.3
# text-unidecode==1.3
# tzdata==2025.2
# urllib3==2.5.0

# ── Django Core ────────────────────────────────────────────────────────────
Django>=4.2,<5.0
pillow>=10.0.0

# ── PDF Extraction ─────────────────────────────────────────────────────────
PyMuPDF>=1.23.0
pytesseract>=0.3.10

# ── NLP ────────────────────────────────────────────────────────────────────
spacy>=3.7.0
sentence-transformers>=2.6.0

# ── Skill Matching ─────────────────────────────────────────────────────────
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0

# ── Scoring ────────────────────────────────────────────────────────────────
scikit-learn>=1.4.0
numpy>=1.26.0